ssh-key: move `AlgorithmName` to a separate module.
